boot-deploy-functions: parameterize omap kindle header size
amazon-soho uses a different certificate size than the rest of the bowser family,
make this a parameter we can set.

For reference:
  - 1024: tate, jem, otter2
  - 848: soho

+++ boot-deploy-functions.sh
@@ -8,6 +8,7 @@ set -eu
 deviceinfo_append_dtb=""
 deviceinfo_arch=""
 deviceinfo_bootimg_amazon_omap_header=""
+deviceinfo_bootimg_amazon_omap_header_size=""
 deviceinfo_bootimg_append_seandroidenforce=""
 deviceinfo_bootimg_custom_args=""
 deviceinfo_bootimg_blobpack=""
@@ -822,12 +823,19 @@ create_bootimg() {
 		mv "$work_dir/boot.temp" $_bootimg
 	fi
 	if [ "${deviceinfo_bootimg_amazon_omap_header}" = "true" ]; then
+		if [ -z "${deviceinfo_bootimg_amazon_omap_header_size}" ]; then
+			log "ERROR: deviceinfo_bootimg_amazon_omap_header is set,"
+			log "but 'deviceinfo_bootimg_amazon_header_size' is missing."
+			log "Please set the certificate header size used by the device."
+			log "See also: <https://postmarketos.org/deviceinfo>"
+			exit 1
+		fi
 		log_arrow "initramfs: prepending Amazon OMAP signature exploit to boot.img"
 		# cert struct start offset and length, used to stack smash cert verification.
 		# returns to address placed in system partition by the cyanogen install,
 		# which jumps to the 2nd stage u-boot at the top of the boot partition.
 		local _header_data='\x50\x03\x00\x00\x00\x25\xE4\x00'
-		local _header_size=$((0x400))
+		local _header_size="${deviceinfo_bootimg_amazon_omap_header_size}"
 		local _header_offset=$((0x34))
 		local _tmp_file="${work_dir}/boot.temp"
 		dd if=/dev/zero bs="${_header_size}" count=1 status=none > "${_tmp_file}"
